Add Image service for handling images as data
We cannot pass the images back as a result of the function call since
that would be too large and the model doesn't have a great way to handle.

So we store the image in local cache and produce a URI that the can be
used to fetch it.  The model can then generate markdown with a link to
this URI.

This is better than using the URIs generated by the model since those
include access tokens.
It's also the only mechanism supported by some models  (gpt-image-1).